chore: 서비스 인터페이스 삭제, 단일 클래스로 변경
